# docs/openapi/taakhelden-core-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wispel core API",
    "version": "0.1.0",
    "description": "Generated core contract snapshot from packages/shared. Covers web dashboard and iOS Phase 1 endpoints."
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "paths": {
...
    "/members/{id}/pause": {
      "get": {
        "summary": "Read active Rustschild pause for a child",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildPause"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Set or replace a per-child pause (parent full)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetChildPauseBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildPause"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/members/children": {
      "post": {
        "summary": "Create a child profile",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberView"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChildBody"
              }
            }
          }
        }
      }
    },
...
    "/members/{id}/avatar": {
      "get": {
        "summary": "Read equipped + unlocked avatar state for a child",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberAvatarState"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Equip unlocked avatar items (idempotent)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipAvatarBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberAvatarState"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "$ref": "#/definitions/ApiError",
        "definitions": {
          "ApiError": {
            "type": "object",
            "properties": {
              "error": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "enum": [
                      "VALIDATION_FAILED",
                      "UNAUTHORIZED",
                      "FORBIDDEN",
                      "NOT_FOUND",
                      "TASK_ALREADY_COMPLETED",
                      "INVALID_STATUS",
                      "UNDO_WINDOW_EXPIRED",
                      "EMAIL_IN_USE",
                      "INVALID_CREDENTIALS",
                      "INSUFFICIENT_POINTS",
                      "REWARD_LIMIT_REACHED",
                      "PIN_LOCKED",
                      "INVALID_FAMILY_CODE",
                      "INVALID_INVITE",
                      "INSTANCE_SLOT_TAKEN",
                      "IDEMPOTENCY_REPLAY",
                      "IDEMPOTENCY_KEY_REUSED",
                      "RATE_LIMITED",
                      "UPSTREAM_UNAVAILABLE"
                    ]
                  },
                  "message": {
                    "type": "string"
                  },
                  "details": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "code",
                  "message"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "error"
            ],
            "additionalProperties": false
          }
        }
      },
...
      "CreateChildBody": {
        "$ref": "#/definitions/CreateChildBody",
        "definitions": {
          "CreateChildBody": {
            "type": "object",
            "properties": {
              "displayName": {
                "type": "string",
                "minLength": 1,
                "maxLength": 30
              },
              "birthYear": {
                "type": "integer",
                "minimum": 2005,
                "maximum": 2023
              },
              "avatarId": {
                "type": "string"
              },
              "pincode": {
                "type": "string",
                "pattern": "^\\d{4}$"
              }
            },
            "required": [
              "displayName",
              "birthYear",
              "pincode"
            ],
            "additionalProperties": false
          }
        }
      },
      "MemberView": {
        "$ref": "#/definitions/MemberView",
        "definitions": {
          "MemberView": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "role": {
                "type": "string",
                "enum": [
                  "parent",
                  "child"
                ]
              },
              "displayName": {
                "type": "string"
              },
              "avatarId": {
                "type": "string",
                "nullable": true
              },
              "ageMode": {
                "type": "string",
                "enum": [
                  "young",
                  "mid",
                  "teen"
                ]
              },
              "birthYear": {
                "type": "integer"
              }
            },
            "required": [
              "id",
              "role",
              "displayName",
              "avatarId",
              "ageMode"
            ],
            "additionalProperties": false
          }
        }
      },
...
      "MemberAvatarState": {
        "$ref": "#/definitions/MemberAvatarState",
        "definitions": {
          "MemberAvatarState": {
            "type": "object",
            "properties": {
              "memberId": {
                "type": "string"
              },
              "equipped": {
                "type": "object",
                "properties": {
                  "hat": {
                    "type": "string",
                    "nullable": true
                  },
                  "background": {
                    "type": "string",
                    "nullable": true
                  },
                  "accessory": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "hat",
                  "background",
                  "accessory"
                ],
                "additionalProperties": false
              },
              "unlocked": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "level": {
                "type": "integer",
                "minimum": 1
              },
              "lifetimeEarned": {
                "type": "integer",
                "minimum": 0
              }
            },
            "required": [
              "memberId",
              "equipped",
              "unlocked",
              "level",
              "lifetimeEarned"
            ],
            "additionalProperties": false
          }
        }
      },
      "EquipAvatarBody": {
        "$ref": "#/definitions/EquipAvatarBody",
        "definitions": {
          "EquipAvatarBody": {
            "type": "object",
            "properties": {
              "hat": {
                "type": "string",
                "nullable": true
              },
              "background": {
                "type": "string",
                "nullable": true
              },
              "accessory": {
                "type": "string",
                "nullable": true
              }
            },
            "additionalProperties": false
          }
        }
      },
...
      "ChildPause": {
        "$ref": "#/definitions/ChildPause",
        "definitions": {
          "ChildPause": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "childId": {
                "type": "string"
              },
              "startsOn": {
                "type": "string"
              },
              "endsOn": {
                "type": "string",
                "nullable": true
              },
              "reason": {
                "type": "string",
                "maxLength": 140,
                "nullable": true
              },
              "active": {
                "type": "boolean"
              }
            },
            "required": [
              "id",
              "childId",
              "startsOn",
              "endsOn",
              "reason",
              "active"
            ],
            "additionalProperties": false
          }
        }
      },
      "SetChildPauseBody": {
        "$ref": "#/definitions/SetChildPauseBody",
        "definitions": {
          "SetChildPauseBody": {
            "type": "object",
            "properties": {
              "startsOn": {
                "type": "string",
                "format": "date"
              },
              "endsOn": {
                "type": "string",
                "format": "date",
                "nullable": true
              },
              "reason": {
                "type": "string",
                "maxLength": 140
              }
            },
            "required": [
              "startsOn"
            ],
            "additionalProperties": false
          }
        }
      },
...
    }
  }
}